// src/app/shared/models/model.ts
export interface url {
  url: string;
}

export interface pageSelection {
  skip: number;
  limit: number;
}

export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface studentRefferal {
  sNo?:number;
  img:string;
  referredId:string;
  referrals:string;
  url:string;
  visits:string;
  totalEarned:string;
}